+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 ## Unreleased
 
 - Storing connection information to the ProPublica API
-- Fetching a list of all members of a chamber in a particular session of Congress
-- Fetching detailed information about one member 
+- Fetching a list of all members of a chamber in a particular session of Congress, or all current members from a state
+- Fetching detailed information about one member